Serialization Function Relocation
